backend/requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
fastapi>=0.68.0,<0.69.0
pydantic==1.10.2
uvicorn>=0.15.0,<0.16.0
# bson==0.5.10
pymongo==4.3.3
connection==2021.7.20
schematics==2.1.1
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.5
motor==3.1.1